refactor(ui): split app sidebar responsibilities
