+++ ui/src/components/app-sidebar-base.ts
@@ -1,0 +1,59 @@
+import { consume } from "@lit/context";
+import { property } from "lit/decorators.js";
+import type { UpdateAvailable } from "../api/types.ts";
+import {
+  DEFAULT_SIDEBAR_PINNED_ROUTES,
+  type NavigationRouteId,
+  type SidebarNavRoute,
+} from "../app-navigation.ts";
+import type { RouteId } from "../app-route-paths.ts";
+import {
+  applicationContext,
+  type ApplicationContext,
+  type ApplicationNavigationOptions,
+} from "../app/context.ts";
+import type { CatalogOpenTarget } from "../app/settings.ts";
+import type { ThemeMode } from "../app/theme.ts";
+import { OpenClawLightDomContentsElement } from "../lit/openclaw-element.ts";
+import type { NewSessionTarget } from "../pages/new-session/location.ts";
+
+/** Stable custom-element inputs. Behavior is layered in focused sidebar modules. */
+export abstract class AppSidebarBase extends OpenClawLightDomContentsElement {
+  @property({ attribute: false }) basePath = "";
+  @property({ attribute: false }) activeRouteId?: NavigationRouteId;
+  @property({ attribute: false }) activePluginTabId = "";
+  @property({ attribute: false }) enabledRouteIds?: readonly NavigationRouteId[];
+  @property({ attribute: false }) connected = false;
+  @property({ attribute: false }) terminalAvailable = false;
+  @property({ attribute: false }) catalogOpenTarget: CatalogOpenTarget = "viewer";
+  @property({ attribute: false }) canPairDevice = false;
+  @property({ attribute: false }) sessionKey = "";
+  @property({ attribute: false }) sidebarPinnedRoutes: readonly SidebarNavRoute[] =
+    DEFAULT_SIDEBAR_PINNED_ROUTES;
+  /** Agents surfaced first in the chip quick switcher when many exist. */
+  @property({ attribute: false }) pinnedAgentIds: readonly string[] = [];
+  @property({ attribute: false }) themeMode: ThemeMode = "system";
+  @property({ attribute: false }) lobsterPetVisits = true;
+  @property({ attribute: false }) lobsterPetSounds = false;
+  @property({ attribute: false }) gatewayVersion: string | null = null;
+  @property({ attribute: false }) devGitBranch: string | null = null;
+  @property({ attribute: false }) updateAvailable: UpdateAvailable | null = null;
+  @property({ attribute: false }) updateRunning = false;
+  @property({ attribute: false }) onUpdate: () => void = () => undefined;
+  @property({ attribute: false }) onOpenPalette?: () => void;
+  @property({ attribute: false }) onToggleSidebar?: () => void;
+  @property({ attribute: false }) onOpenNewSession?: (
+    agentId: string,
+    target?: NewSessionTarget,
+  ) => void;
+  /** Agent id of the in-flight new-session draft; renders the draft row. */
+  @property({ attribute: false }) draftSessionAgentId = "";
+  @property({ attribute: false }) onUpdatePinnedRoutes?: (routes: SidebarNavRoute[]) => void;
+  @property({ attribute: false }) onPairMobile?: () => void;
+  @property({ attribute: false })
+  onNavigate?: (routeId: NavigationRouteId, options?: ApplicationNavigationOptions) => void;
+  @property({ attribute: false }) onPreloadRoute?: (routeId: NavigationRouteId) => Promise<void>;
+
+  @consume({ context: applicationContext, subscribe: true })
+  protected context?: ApplicationContext<RouteId>;
+}
